+++ src/index.ts
@@ -1,6 +1,8 @@
 export { default as append } from './append';
 export { default as insert } from './insert';
 export { default as remove } from './remove';
+export { default as removeMatch } from './removeMatch';
 export { default as set } from './set';
 export { default as swap } from './swap';
 export { default as update } from './update';
+export { default as updateMatch } from './updateMatch';