+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.1.0 (2023-12-13)
+
+- Add: `removeMatch()` removes some items matching the given predicate. It's a
+  combination of `[].findIndex()` and `remove()` or `[].filter()` with inversion
+  and limit.
+- Add: `updateMatch()` updates some items matching the given predicate. It's a
+  combination of `[].map()` with limit.
+
 ## 1.0.0 (2022-06-25)
 
 - Initial release.